@@ -9,17 +9,17 @@
         }
     ],
     "require": {
-        "php": ">=5.6.4",
-        "illuminate/validation": "5.4.*",
-        "illuminate/events": "5.4.*",
-        "illuminate/routing": "5.4.*",
-        "illuminate/contracts": "5.4.*",
-        "illuminate/support": "5.4.*",
+        "php": ">=7.0.0",
+        "illuminate/validation": "5.5.*",
+        "illuminate/events": "5.5.*",
+        "illuminate/routing": "5.5.*",
+        "illuminate/contracts": "5.5.*",
+        "illuminate/support": "5.5.*",
         "congraph/contracts": "dev-laravel8"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.7",
-        "orchestra/testbench": "3.4.*"
+        "phpunit/phpunit": "~6.0",
+        "orchestra/testbench": "3.5.*"
     },
     "autoload": {
         "psr-4": {
